compilation warnings

+++ src/qsm/QSMbuilder_radius.cpp
@@ -54,7 +54,8 @@ public:
 void QSMbuilder::construct_radii(const PointCloud& tree, double tip_radius)
 {
   double H = 0.0;
-  for (int i = 0; i < tree.size(); ++i) {
+  for (size_t i = 0; i < tree.size(); ++i)
+  {
     double z = tree.get_z(i);
     if (z > H) H = z;
   }
